% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_zones.R
\name{convert_to_local_time}
\alias{convert_to_local_time}
\title{Convert the \code{POSIXct} columns of a data frame to the local time zone}
\usage{
convert_to_local_time(.data, tz = "")
}
\arguments{
\item{.data}{A tibble with \code{POSIXct} columns}

\item{tz}{The time zone to convert to. The default is the local time zone.}
}
\description{
Tiingo returns data with a UTC time zone. Often you will want to view this
in your own time zone. This function converts each \code{POSIXct} column of the
returned \code{tibble} to the local (or specified) time zone.
}
