% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riingo-package.R
\docType{package}
\name{riingo-package}
\alias{riingo}
\alias{riingo-package}
\title{riingo: An R Interface to the 'Tiingo' Stock Price API}
\description{
Functionality to download stock prices,
    cryptocurrency data, and more from the 'Tiingo' API
    <https://api.tiingo.com/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/business-science/riingo}
  \item Report bugs at \url{https://github.com/business-science/riingo/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Authors:
\itemize{
  \item Matt Dancho \email{mdancho@business-science.io}
}

}
\keyword{internal}
