% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{riingo_crypto_meta}
\alias{riingo_crypto_meta}
\title{Get meta data about a cryptocurrency on Tiingo}
\usage{
riingo_crypto_meta(ticker)
}
\arguments{
\item{ticker}{One or more cryptocurrency tickers.
Specified as \code{"btcusd"} for bitcoin quoted in USD. A character vector.}
}
\description{
Relevant returned meta data include: ticker, name, description, quote currency,
and base currency.
}
\examples{

\donttest{

# Bitcoin meta
riingo_crypto_meta("btcusd")

# A trick to return ALL crypto meta data
# For some reason Descriptions are not returned here
riingo_crypto_meta("")

}

}
