% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals-definitions.R
\name{riingo_fundamentals_definitions}
\alias{riingo_fundamentals_definitions}
\title{Fundamentals - Definitions}
\usage{
riingo_fundamentals_definitions()
}
\value{
A data frame with columns describing the various data codes that
will be returned by \code{\link[=riingo_fundamentals_statements]{riingo_fundamentals_statements()}}.
}
\description{
This function can be used to check which fields are available from the
fundamentals endpoint. As Tiingo adds more indicators, the output of this
function may change to reflect the addition of new indicators.
}
\examples{
\dontrun{
riingo_fundamentals_definitions()
}
}
\seealso{
Other fundamentals: 
\code{\link{riingo_fundamentals_meta}()},
\code{\link{riingo_fundamentals_metrics}()},
\code{\link{riingo_fundamentals_statements}()}
}
\concept{fundamentals}
