% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals-meta.R
\name{riingo_fundamentals_meta}
\alias{riingo_fundamentals_meta}
\title{Fundamentals - Meta}
\usage{
riingo_fundamentals_meta(ticker)
}
\arguments{
\item{ticker}{One or more tickers to download meta information for.}
}
\value{
A data frame with one row per ticker containing fundamentals meta data.
}
\description{
This function can be used to get fundamentals meta data for individual
tickers, which includes information about when that ticker was last updated
with new fundamentals data, along with things such as the sector and
industry that that company belongs to.
}
\examples{
\dontrun{
riingo_fundamentals_meta(c("AAPL", "MSFT"))
}
}
\seealso{
Other fundamentals: 
\code{\link{riingo_fundamentals_definitions}()},
\code{\link{riingo_fundamentals_metrics}()},
\code{\link{riingo_fundamentals_statements}()}
}
\concept{fundamentals}
