% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{riingo_iex_quote}
\alias{riingo_iex_quote}
\title{Quote and Top of Book data for a given ticker}
\usage{
riingo_iex_quote(ticker)
}
\arguments{
\item{ticker}{One or more tickers to download data for from Tiingo. Can be a
stock, mutual fund, or ETF. A character vector.}
}
\description{
Tiingo is plugged into the IEX feed, and they provide last sale data along
with TOP (top of book) bid and ask quotes. Note that this cannot be historically
queried.
}
\details{
At the end of the day, the \code{mid}, \code{askPrice}, \code{bidSize}, \code{bidPrice}, \code{askSize},
and \code{lastSize} fields will be \code{NA}. This is normal.
}
\examples{

riingo_iex_quote("QQQ")

}
