% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{riingo_latest}
\alias{riingo_latest}
\title{The latest day's worth of data for a given ticker}
\usage{
riingo_latest(ticker)
}
\arguments{
\item{ticker}{One or more tickers to download data for from Tiingo. Can be a
stock, mutual fund, or ETF. A character vector.}
}
\description{
This returns only the most recent day of daily data for the supplied ticker(s).
}
\examples{

The latest available day of daily data for QQQ
riingo_latest("QQQ")

}
