% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{riingo_latest_iex}
\alias{riingo_latest_iex}
\title{The latest day's worth of intraday data for a given ticker}
\usage{
riingo_latest_iex(ticker, resample_frequency = "1min")
}
\arguments{
\item{ticker}{One or more tickers to download data for from Tiingo. Can be a
stock, mutual fund, or ETF. A character vector.}

\item{resample_frequency}{For Tiingo data, a character specified as one of:
\code{"daily"}, \code{"monthly"}, \code{"quarterly"} or \code{"yearly"}. For IEX data, a character
specified at the \code{"min"} or \code{"hour"} frequencies in the form:
\code{"1min"}, \code{"5min"}, or \code{"2hour"}.}
}
\description{
This returns only the most recent day of intraday data for the supplied ticker(s).
}
\examples{

The latest available day of intraday data for QQQ
riingo_latest_iex("QQQ")

riingo_latest_iex("QQQ", "1hour")

}
