% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{riingo_meta}
\alias{riingo_meta}
\title{Get meta data about a ticker available on Tiingo}
\usage{
riingo_meta(ticker)
}
\arguments{
\item{ticker}{One or more tickers to download data for from Tiingo. Can be a
stock, mutual fund, or ETF. A character vector.}
}
\description{
Retrieve start and end dates for available ticker data, along with the name,
exchange, and description of the ticker.
}
\examples{

\donttest{

riingo_meta("AAPL")

riingo_meta("QQQ")

}

}
