% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{riingo_news}
\alias{riingo_news}
\title{Get Tiingo news feed for a single or series of tickers}
\usage{
riingo_news(ticker, start_date = NULL, end_date = NULL, tags = NULL,
  source = NULL, limit = 100, offset = 0)
}
\arguments{
\item{ticker}{One or more tickers to download data for from Tiingo. Can be a
stock, mutual fund, or ETF. A character vector.}

\item{start_date}{The first date to download data for.
A character in the form YYYY-MM-DD, or a \code{Date} variable.}

\item{end_date}{The last date to download data for.
A character in the form YYYY-MM-DD, or a \code{Date} variable.}

\item{tags}{Tags Requested. A character vector}

\item{source}{Requested news source URL.}

\item{limit}{The maximum number of articles to be retrieved. The default is 100, the maxiumum is 1,000}

\item{offset}{Pagination variable used alongside limit. Returns an array wih the results shifted by this parameter}
}
\description{
Get Tiingo news feed for a single or series of tickers
}
\examples{

\dontrun{

riingo_news("QQQ")

}

}
