% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker_info.R
\name{is_supported_ticker}
\alias{is_supported_ticker}
\alias{supported_tickers}
\title{Ticker information}
\usage{
is_supported_ticker(ticker)

supported_tickers()
}
\arguments{
\item{ticker}{The single ticker to check for on Tiingo.}
}
\description{
\code{\link[=is_supported_ticker]{is_supported_ticker()}} can tell you if a given ticker is supported on Tiingo.
\code{\link[=supported_tickers]{supported_tickers()}} returns a \code{tibble} listing every available ticker.
}
